cli: rename cylc_ script modules
