+++ cylc/flow/scripts/extract_resources.py
