+++ cylc/flow/scripts/get_directory.py
