+++ cylc/flow/scripts/register.py
