+++ cylc/flow/scripts/remove.py
