+++ cylc/flow/scripts/show.py
