+++ setup.cfg
@@ -67,57 +67,57 @@ console_scripts =
     cylc = cylc.flow.scripts.cylc:main
 # cylc subcommands
 cylc.command =
-    broadcast = cylc.flow.scripts.cylc_broadcast:main
-    cat-log = cylc.flow.scripts.cylc_cat_log:main
-    check-versions = cylc.flow.scripts.cylc_check_versions:main
-    checkpoint = cylc.flow.scripts.cylc_checkpoint:main
-    client = cylc.flow.scripts.cylc_client:main
-    cycle-point = cylc.flow.scripts.cylc_cycle_point:main
-    diff = cylc.flow.scripts.cylc_diff:main
-    dump = cylc.flow.scripts.cylc_dump:main
-    edit = cylc.flow.scripts.cylc_edit:main
-    ext-trigger = cylc.flow.scripts.cylc_ext_trigger:main
-    extract-resources = cylc.flow.scripts.cylc_extract_resources:main
-    function-run = cylc.flow.scripts.cylc_function_run:main
-    get-directory = cylc.flow.scripts.cylc_get_directory:main
-    get-site-config = cylc.flow.scripts.cylc_get_site_config:main
-    get-suite-config = cylc.flow.scripts.cylc_get_suite_config:main
-    get-suite-contact = cylc.flow.scripts.cylc_get_suite_contact:main
-    get-suite-version = cylc.flow.scripts.cylc_get_suite_version:main
-    graph = cylc.flow.scripts.cylc_graph:main
-    hold = cylc.flow.scripts.cylc_hold:main
-    jobs-kill = cylc.flow.scripts.cylc_jobs_kill:main
-    jobs-poll = cylc.flow.scripts.cylc_jobs_poll:main
-    jobs-submit = cylc.flow.scripts.cylc_jobs_submit:main
-    kill = cylc.flow.scripts.cylc_kill:main
-    list = cylc.flow.scripts.cylc_list:main
-    ls-checkpoints = cylc.flow.scripts.cylc_ls_checkpoints:main
-    message = cylc.flow.scripts.cylc_message:main
-    nudge = cylc.flow.scripts.cylc_nudge:main
-    ping = cylc.flow.scripts.cylc_ping:main
-    poll = cylc.flow.scripts.cylc_poll:main
-    psutils = cylc.flow.scripts.cylc_psutil:main
-    register = cylc.flow.scripts.cylc_register:main
-    release = cylc.flow.scripts.cylc_release:main
-    reload = cylc.flow.scripts.cylc_reload:main
-    remote-init = cylc.flow.scripts.cylc_remote_init:main
-    remote-tidy = cylc.flow.scripts.cylc_remote_tidy:main
-    remove = cylc.flow.scripts.cylc_remove:main
-    report-timings = cylc.flow.scripts.cylc_report_timings:main
-    restart = cylc.flow.scripts.cylc_restart:main
-    run = cylc.flow.scripts.cylc_run:main
-    scan = cylc.flow.scripts.cylc_scan:cli
-    search = cylc.flow.scripts.cylc_search:main
-    set-verbosity = cylc.flow.scripts.cylc_set_verbosity:main
-    show = cylc.flow.scripts.cylc_show:main
-    set-outputs = cylc.flow.scripts.cylc_set_outputs:main
-    stop = cylc.flow.scripts.cylc_stop:main
-    subscribe = cylc.flow.scripts.cylc_subscribe:main
-    suite-state = cylc.flow.scripts.cylc_suite_state:main
-    tui = cylc.flow.scripts.cylc_tui:main
-    trigger = cylc.flow.scripts.cylc_trigger:main
-    validate = cylc.flow.scripts.cylc_validate:main
-    view = cylc.flow.scripts.cylc_view:main
+    broadcast = cylc.flow.scripts.broadcast:main
+    cat-log = cylc.flow.scripts.cat_log:main
+    check-versions = cylc.flow.scripts.check_versions:main
+    checkpoint = cylc.flow.scripts.checkpoint:main
+    client = cylc.flow.scripts.client:main
+    cycle-point = cylc.flow.scripts.cycle_point:main
+    diff = cylc.flow.scripts.diff:main
+    dump = cylc.flow.scripts.dump:main
+    edit = cylc.flow.scripts.edit:main
+    ext-trigger = cylc.flow.scripts.ext_trigger:main
+    extract-resources = cylc.flow.scripts.extract_resources:main
+    function-run = cylc.flow.scripts.function_run:main
+    get-directory = cylc.flow.scripts.get_directory:main
+    get-site-config = cylc.flow.scripts.get_site_config:main
+    get-suite-config = cylc.flow.scripts.get_suite_config:main
+    get-suite-contact = cylc.flow.scripts.get_suite_contact:main
+    get-suite-version = cylc.flow.scripts.get_suite_version:main
+    graph = cylc.flow.scripts.graph:main
+    hold = cylc.flow.scripts.hold:main
+    jobs-kill = cylc.flow.scripts.jobs_kill:main
+    jobs-poll = cylc.flow.scripts.jobs_poll:main
+    jobs-submit = cylc.flow.scripts.jobs_submit:main
+    kill = cylc.flow.scripts.kill:main
+    list = cylc.flow.scripts.list:main
+    ls-checkpoints = cylc.flow.scripts.ls_checkpoints:main
+    message = cylc.flow.scripts.message:main
+    nudge = cylc.flow.scripts.nudge:main
+    ping = cylc.flow.scripts.ping:main
+    poll = cylc.flow.scripts.poll:main
+    psutils = cylc.flow.scripts.psutil:main
+    register = cylc.flow.scripts.register:main
+    release = cylc.flow.scripts.release:main
+    reload = cylc.flow.scripts.reload:main
+    remote-init = cylc.flow.scripts.remote_init:main
+    remote-tidy = cylc.flow.scripts.remote_tidy:main
+    remove = cylc.flow.scripts.remove:main
+    report-timings = cylc.flow.scripts.report_timings:main
+    restart = cylc.flow.scripts.restart:main
+    run = cylc.flow.scripts.run:main
+    scan = cylc.flow.scripts.scan:cli
+    search = cylc.flow.scripts.search:main
+    set-verbosity = cylc.flow.scripts.set_verbosity:main
+    show = cylc.flow.scripts.show:main
+    set-outputs = cylc.flow.scripts.set_outputs:main
+    stop = cylc.flow.scripts.stop:main
+    subscribe = cylc.flow.scripts.subscribe:main
+    suite-state = cylc.flow.scripts.suite_state:main
+    tui = cylc.flow.scripts.tui:main
+    trigger = cylc.flow.scripts.trigger:main
+    validate = cylc.flow.scripts.validate:main
+    view = cylc.flow.scripts.view:main
 # async functions to run within the scheduler main loop
 cylc.main_loop =
     health_check = cylc.flow.main_loop.health_check
